create circle dom test factories
